Board 0 and 1 project build. MAVLink RC channel values.
