+++ pico/ctrl/.vscode/settings.json
@@ -35,6 +35,8 @@
         "NOBLOCK",
         "nvic",
         "OLAT",
+        "pbsw",
+        "picoprobe",
         "picotool",
         "pindirs",
         "piosm",
@@ -64,6 +66,7 @@
         "SRXL",
         "stdlib",
         "SYSIRQ",
+        "TERMX",
         "THRSH",
         "TXFIFO",
         "txrx",
