+++ pico/ctrl/src/CMakeLists.txt
@@ -31,12 +31,10 @@ add_compile_options(
 )
 pico_sdk_init()
 
-# Local libraries (our additional sources)
-add_subdirectory(cmt)
+
 add_subdirectory(curswitch)
 add_subdirectory(dcs)
 add_subdirectory(display)
-add_subdirectory(expio)
 add_subdirectory(gfx)
 add_subdirectory(hid)
 add_subdirectory(hwrt)
@@ -47,14 +45,13 @@ add_subdirectory(rover)
 add_subdirectory(sensbank)
 add_subdirectory(servo)
 add_subdirectory(term)
+add_subdirectory(termx)
 add_subdirectory(touch_panel)
-add_subdirectory(util)
-#
-# External libraries
-add_subdirectory(lib)
 
-# Add executable.
-add_executable(hwctrl
+#
+# Common top-level sources
+#
+set(CMN_TL_SRC
   main.c
   board.c
   debug_support.c
@@ -64,17 +61,132 @@ add_executable(hwctrl
   tests.c
   p2bugs.c
 )
+#
+# Common libraries (built the same for both boards 0 & 1)
+#
+add_subdirectory(cmt)
+add_subdirectory(expio)
+add_subdirectory(util)
+#
+# External libraries
+add_subdirectory(lib)
+#
+set(CMN_LIBS
+  cmt
+  expio
+  sensbank
+  util
+  hardware_adc
+  hardware_clocks
+  hardware_dma
+  hardware_exception
+  hardware_i2c
+  hardware_pio
+  hardware_spi
+  hardware_timer
+  pico_double
+  pico_float
+  pico_multicore
+  pico_stdlib
+)
 
-pico_set_program_name(hwctrl "HWControl")
-pico_set_program_version(hwctrl "0.1")
-
+#
+# Add executable for Board-0
+#
+add_executable(hwctrl_0
+  ${CMN_TL_SRC}
+)
+pico_set_program_name(hwctrl_0 "HWControl-0")
+pico_set_program_version(hwctrl_0 "1.0")
 # Use the VFP instructions for float and double (can use DCP)
-pico_set_float_implementation(hwctrl pico)
-pico_set_double_implementation(hwctrl pico)
-
+pico_set_float_implementation(hwctrl_0 pico)
+pico_set_double_implementation(hwctrl_0 pico)
 # Use the USB for stdio. UART is used to communicate with host for commands.
-pico_enable_stdio_uart(hwctrl 1)
-pico_enable_stdio_usb(hwctrl 0)
+pico_enable_stdio_uart(hwctrl_0 1)
+pico_enable_stdio_usb(hwctrl_0 0)
+target_compile_definitions(hwctrl_0 PRIVATE
+	BOARD_ADDR=0
+)
+target_include_directories(hwctrl_0 PRIVATE
+	${CMAKE_CURRENT_LIST_DIR}
+)
+# Add the libraries required by the system to the build
+target_link_libraries(hwctrl_0
+  ${CMN_LIBS}
+)
+target_link_libraries(hwctrl_0
+  dcs
+  hwrt0
+  rcrx
+  rover
+  servo
+  termx
+)
+if(${PICO_BOARD} STREQUAL "pico")
+target_link_libraries(hwctrl_0
+  hardware_rtc
+)
+endif()
+if(${PICO_BOARD} STREQUAL "pico_w")
+target_link_libraries(hwctrl_0
+  hardware_rtc
+  pico_cyw43_arch_lwip_threadsafe_background
+)
+endif()
+#
+pico_add_extra_outputs(hwctrl_0)
+
+
+#
+# Add executable for Board-1
+#
+add_executable(hwctrl_1
+  ${CMN_TL_SRC}
+)
+pico_set_program_name(hwctrl_1 "HWControl-1")
+pico_set_program_version(hwctrl_1 "1.0")
+# Use the VFP instructions for float and double (can use DCP)
+pico_set_float_implementation(hwctrl_1 pico)
+pico_set_double_implementation(hwctrl_1 pico)
+# Use the USB for stdio. UART is used to communicate with host for commands.
+pico_enable_stdio_uart(hwctrl_1 0)
+pico_enable_stdio_usb(hwctrl_1 1)
+target_compile_definitions(hwctrl_1 PRIVATE
+	BOARD_ADDR=1
+)
+target_include_directories(hwctrl_1 PRIVATE
+	${CMAKE_CURRENT_LIST_DIR}
+)
+# Add the libraries required by the system to the build
+target_link_libraries(hwctrl_1
+  ${CMN_LIBS}
+)
+target_link_libraries(hwctrl_1
+  curswitch
+  display
+  fonts
+  gfx
+  hid
+  hwrt1
+  neopix
+  rotary_encoder
+  terminal
+  termx
+  touch_panel
+)
+if(${PICO_BOARD} STREQUAL "pico")
+target_link_libraries(hwctrl_1
+  hardware_rtc
+)
+endif()
+if(${PICO_BOARD} STREQUAL "pico_w")
+target_link_libraries(hwctrl_1
+  hardware_rtc
+  pico_cyw43_arch_lwip_threadsafe_background
+)
+endif()
+#
+pico_add_extra_outputs(hwctrl_1)
 
 # Add the required include file paths to the build
 include_directories(
@@ -85,10 +197,21 @@ include_directories(
   ${PICO_SDK_PATH}/src/rp2_common/hardware_i2c/include
   ${PICO_SDK_PATH}/src/rp2_common/hardware_pio/include
   ${PICO_SDK_PATH}/src/rp2_common/hardware_pwm/include
-  ${PICO_SDK_PATH}/src/rp2_common/hardware_rtc/include
-  ${PICO_SDK_PATH}/src/rp2_common/pico_cyw43_arch/include
   ${PICO_SDK_PATH}/src/rp2_common/pico_multicore/include
 )
+
+if(${PICO_BOARD} STREQUAL "pico")
+include_directories(
+  ${PICO_SDK_PATH}/src/rp2_common/hardware_rtc/include
+)
+endif()
+if(${PICO_BOARD} STREQUAL "pico_w")
+include_directories(
+  ${PICO_SDK_PATH}/src/rp2_common/hardware_rtc/include
+  ${PICO_SDK_PATH}/src/rp2_common/pico_cyw43_arch/include
+)
+endif()
+
 
 # ZZZ Figure out how to set `PICO_DEBUG_MALLOC` only if 'debug build'
 add_compile_definitions(
@@ -105,51 +228,5 @@ add_compile_definitions(
   PICO_STDIO_USB_SUPPORT_CHARS_AVAILABLE_CALLBACK
 )
 
-# Add the libraries required by the system to the build
-target_link_libraries(hwctrl
-  cmt
-  curswitch
-  dcs
-  display
-  expio
-  fonts
-  gfx
-  hid
-  hwrt
-  neopix
-  rcrx
-  rotary_encoder
-  rover
-  sensbank
-  servo
-  terminal
-  touch_panel
-  util
-  hardware_adc
-  hardware_clocks
-  hardware_dma
-  hardware_exception
-  hardware_i2c
-  hardware_pio
-  hardware_spi
-  hardware_timer
-  pico_double
-  pico_float
-  pico_multicore
-  pico_stdlib
-)
-if(${PICO_BOARD} STREQUAL "pico")
-target_link_libraries(hwctrl
-  hardware_rtc
-)
-endif()
-if(${PICO_BOARD} STREQUAL "pico_w")
-target_link_libraries(hwctrl
-  hardware_rtc
-  pico_cyw43_arch_lwip_threadsafe_background
-)
-endif()
 
 # pico_set_linker_script(hwctrl ${CMAKE_SOURCE_DIR}/memmap_custom.ld)
-
-pico_add_extra_outputs(hwctrl)
