+++ pico/ctrl/src/board.h
@@ -23,6 +23,7 @@ extern "C" {
 #endif
 
 #include <stdbool.h>
+#include <stdint.h>
 
 /**
  * @brief Initialize the board
