+++ pico/ctrl/src/hwrt/CMakeLists.txt
@@ -1,12 +1,21 @@
-# Library: Hardware Runtime (hwrt) (obj only)
-add_library(hwrt INTERFACE)
+# Library: Hardware Runtime for board address 0 (hwrt0) (obj only)
+add_library(hwrt0 INTERFACE)
 
-target_sources(hwrt INTERFACE
-    hwrt.c
+target_sources(hwrt0 INTERFACE
+    hwrt0.c
 )
 
-#add_subdirectory(display)
+target_link_libraries(hwrt0 INTERFACE
+    pico_stdlib
+)
 
-target_link_libraries(hwrt INTERFACE
+# Library: Hardware Runtime for board address 1 (hwrt1) (obj only)
+add_library(hwrt1 INTERFACE)
+
+target_sources(hwrt1 INTERFACE
+    hwrt1.c
+)
+
+target_link_libraries(hwrt1 INTERFACE
     pico_stdlib
 )
