+++ pico/ctrl/src/hwrt/hwrt.h
@@ -1,5 +1,5 @@
 /**
- * HWControl Operating System - Base.
+ * Hardware Runtime.
  *
  * Setup for the message loop and idle processing.
  *
@@ -13,19 +13,17 @@
 extern "C" {
 #endif
 
-#include "cmt/cmt.h"
-
 #define HWRT_CORE_NUM 0
 
 /**
- * @brief Initialize the back-end
- * @ingroup backend
+ * @brief Initialize the runtime
+ * @ingroup hwrt
  */
 extern void hwrt_module_init(void);
 
 /**
- * @brief Start the Backend (core 0 (endless) message-loop).
- * @ingroup backend
+ * @brief Start the runtime (core 0 (endless) message-loop).
+ * @ingroup hwrt
  */
 extern void start_hwrt(void);
