+++ pico/ctrl/src/rotary_encoder/re_pbsw.c
@@ -19,7 +19,7 @@
 
 static void _gpio_event_string(char* buf, uint32_t events);
 
-void re_pbsw_irq_handler(uint gpio, uint32_t events) {
+void re_pbsw_irq_handler(uint32_t gpio, uint32_t events) {
     if (events & GPIO_IRQ_EDGE_FALL) {
         cmt_msg_t msg;
         cmt_msg_init(&msg, MSG_MAIN_USER_SWITCH_PRESS);
