+++ pico/ctrl/src/rotary_encoder/re_pbsw.h
@@ -14,7 +14,7 @@
 extern "C" {
 #endif
 
-extern void re_pbsw_irq_handler(uint gpio, uint32_t events);
+extern void re_pbsw_irq_handler(uint32_t gpio, uint32_t events);
 
 /**
  * @brief Initialize the rotary encoder push-button switch library.
