+++ pico/ctrl/src/rotary_encoder/rotary_encoder.h
@@ -14,6 +14,8 @@
 extern "C" {
 #endif
 
+#include <pico/types.h>
+
 /**
  * @brief Current position count of the rotary encoder.
  *
