+++ pico/ctrl/src/termx/CMakeLists.txt
@@ -1,0 +1,9 @@
+# Library: Touch Panel (obj only)
+add_library(termx INTERFACE)
+
+target_sources(termx INTERFACE
+)
+
+target_link_libraries(termx INTERFACE
+    pico_stdlib
+)
